marche

+++ src/main/java/tn/esprit/spring/kaddem/services/ContratServiceImpl.java
@@ -87,7 +87,7 @@ ContratRepository contratRepository;
 }
 
 	public float getChiffreAffaireEntreDeuxDates(Date startDate, Date endDate){
-		float differenceInTime = (float) (endDate.getTime() - startDate.getTime());
+		float differenceInTime = endDate.getTime() - startDate.getTime();
 		float differenceInDays = (differenceInTime / (1000 * 60 * 60 * 24)) % 365;
 		float differenceInMonths =differenceInDays/30;
         List<Contrat> contrats=contratRepository.findAll();
